feat: Imputation logic and TIFU
Why:
1. To use the raw feature `slam` with lookback time window manner,
the imputation can't be done by merely filling with 0, because the
actual meaning is that the client didn't make a transaction on that
`shop_tag` in the previous month. `slam` still exists implicitly.
2. To help generating features related to TIFU-KNN, core functions
are implemented.

What:
1. Modify imputation logic for raw feature `slam` by imputing the
missing `slam` with `slam` value in the most recent sample.
2. Implement core functions facilitating the generation of features
proposed in TIFU-KNN.