feat: Feature prediction vectors
Why:
The organization and configuration for multiple feature vector
specification is hard to use and prone to error. Hence, redesign
is of vital importance for the further feature engineering and
selection process. Moreover, manual selection on `shop_tag` dim
in each feature vector is enabled in this version.

What:
1. Redesign feature prediction vector specification.
2. Add functionality for manual selection on `shop_tag` dim in each
feature vector.
------------------------------------------------------------------
Other things missed in last version:
1. Modify transaction related feature `n_shop_tags_vec` to have
hard-coded time dimension, because length of dts are different
for training and validation sets.