refactor: Parameter packing and gp stats query
Why:
To make time interval constraint more flexible when computing groupby
stats, the configuration of time_slots is modified. Another issue
is raised when computing stats using `agg` with multiple functions
at once.

What:
1. Modify `time_slots` configuration for groupby stats derivation;
that is, usage of (time_lower, time_upper) is changed to make time
constraint more flexible.
2. Redesign stats function getter to fix the stats derivation error,
which returns the same value for different quantile specifications.
That is, stats derivation returns the same value x for quantile0.25
and quantile0.5, which shouldn't behave like this. And, I still need
to think a better way to solve this.